Refactored handling of connection errors, now able to reconnect even if all IPs change.
Local endpoint, Sonos IP and both changing now allow for recovering the connection. The recovery doesn't happen until the next call and it might not be complete by the next call (especially soap actions). The discovery should not require a restart.

Issues: the topology subscription might not be restored completely.